template_py/pyproject.toml
#   Copyright 2024 Matthew Ralston
#
#   Licensed under the Apache License, Version 2.0 (the "License");
#   you may not use this file except in compliance with the License.
#   You may obtain a copy of the License at
#
#       http://www.apache.org/licenses/LICENSE-2.0
#
#   Unless required by applicable law or agreed to in writing, software
#   distributed under the License is distributed on an "AS IS" BASIS,
#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#   See the License for the specific language governing permissions and
#   limitations under the License.



[build-system]

# pip  
# conda  4.7.12
# wheel 0.43.0
# setuptools 61.3.1
# requires numpy and cython for custom cython correlation function

requires = ["setuptools>=69.2.0", "numpy>=1.21.2", "Cython>=3.0.8"]
build-backend = "setuptools.build_meta"

[project]
name = "template_py"
version = "0.0.2"
description = "Yet another template for Python"
readme = "README.md"
authors = [{name="Matt Ralston <professional.bio.coder@gmail.com>", email="mralston.development@gmail.com"}]
license = { file = "LICENSE.txt" }
classifiers = [
	    "Development Status :: 1 - Planning",
	    "Development Status :: 2 - Pre-Alpha",
	    "Development Status :: 7 - Inactive",
	    "Environment :: Console",
	    "Intended Audience :: Developers",
	    "Intended Audience :: Science/Research",
	    "License :: OSI Approved :: Apache Software License",
	    "Operating System :: OS Independent",
	    "Programming Language :: Python",
	    "Programming Language :: Python :: 3",
	    "Programming Language :: Python :: 3.6",
	    "Programming Language :: Python :: 3.7",
	    "Programming Language :: Python :: 3.8",
	    "Programming Language :: Python :: 3.9",
	    "Programming Language :: Python :: 3.10",
	    "Programming Language :: Python :: 3.12",
	    "Programming Language :: Python :: 3 :: Only",
	    "Operating System :: PalmOS",
	    "Topic :: Scientific/Engineering",
	    "Topic :: Scientific/Engineering :: Bio-Informatics",
	    "Topic :: Software Development :: Libraries :: Python Modules",
	    "Intended Audience :: Other Audience",
	    "License :: Free For Educational Use",
	    "License :: Free for non-commercial use",
	    "License :: OSI Approved :: Academic Free License (AFL)",
	    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
	    "Topic :: File Formats",
	    "Topic :: Games/Entertainment :: Puzzle Games",
	    "Topic :: Text Processing :: Indexing",
	    "Topic :: Text Processing",
	    "Operating System :: POSIX :: Linux",
]

	    #"Topic :: Software Development :: Assemblers"
keywords = ["keyword"]
dependencies = [
	     #"numpy>=2.0.0",
	     "PyYAML>=6.0.1",
	     "jsonschema>=4.17.3",
	     "psutil>=4.2.0",
	     #"Cython>=3.0.8",
	     #"pandas>=2.2.2",
             "setuptools>=69.2.0",
	     ]

#requires-python = ">=3.7.4"
requires-python = ">=3.12.2"

[project.optional-dependencies]
dev = [
    #########################################
    # Build system
    #########################################
    'auditwheel>=5.1.2',
    'build>=0.9.0',
    'coverage>=4.5.4',
    'expects>=0.9.0',
    # 'networkx'
    # 'rypy2>=3.4.2'
    # Vanity
    #'ghstats>=1.2.0',
    #########################################
    # Documentation
    #########################################
    'docutils>=0.17',
    #'grip>=4.6.2',
    'sphinx>=8.0.2',
    #'sphinx-autodoc',
    # These are kind of developmental things, there really cant be testing around a 5% tested codebase, this is open source at its worse,
    # and considering  otherwise is toxic.
    'pytest>=5.3.5',
    #########################################
    #
    # Miss me with it
    #########################################
    'twine==4.0.1'
]


[tool.setuptools]
include-package-data = true
packages = ['template_py']

[tool.setuptools.package-data]
citation = ["CITATION.txt"]


[project.urls]
Homepage = "https://matthewralston.github.io/template_py"
QuickStart = "https://matthewralston.github.io/template_py/quickstart.html"

Github = "https://github.com/MatthewRalston/template_py"
Issues = "https://github.com/MatthewRalston/template_py/issues"
Projects = "https://github.com/MatthewRalston?tab=projects"


[project.scripts]
template_py = "template_py:cli"
#[options.entry_points]



